requirements.txt

pdfminer.six==20220524
